pass api handle in from top
