+++ src/inspect_ai/model/_providers/util/__init__.py
@@ -1,17 +1,19 @@
 from .chatapi import (
+    ChatAPIHandler,
     chat_api_input,
     chat_api_request,
     is_chat_api_rate_limit,
 )
-from .tools import chat_api_tools_handler, parse_tool_call
-from .util import as_stop_reason, model_base_url
+from .llama31 import Llama31Handler
+from .util import as_stop_reason, model_base_url, parse_tool_call
 
 __all__ = [
     "as_stop_reason",
     "chat_api_request",
     "chat_api_input",
-    "chat_api_tools_handler",
     "is_chat_api_rate_limit",
     "model_base_url",
     "parse_tool_call",
+    "ChatAPIHandler",
+    "Llama31Handler",
 ]
