+++ src/inspect_ai/model/_providers/util/util.py
@@ -1,5 +1,12 @@
+import json
 import os
 from logging import getLogger
+from typing import Any
+
+import yaml
+
+from inspect_ai.tool._tool_call import ToolCall
+from inspect_ai.tool._tool_info import ToolInfo
 
 from ..._model_output import StopReason
 
@@ -34,3 +41,52 @@ def model_base_url(base_url: str | None, env_vars: str | list[str]) -> str | Non
             return base_url
 
     return os.getenv("INSPECT_EVAL_MODEL_BASE_URL", None)
+
+
+def parse_tool_call(
+    id: str, function: str, arguments: str, tools: list[ToolInfo]
+) -> ToolCall:
+    error: str | None = None
+    arguments_dict: dict[str, Any] = {}
+
+    def report_parse_error(ex: Exception) -> None:
+        nonlocal error
+        error = f"Error parsing the following tool call arguments:\n\n{arguments}\n\nError details: {ex}"
+        logger.info(error)
+
+    # if the arguments is a dict, then handle it with a plain json.loads
+    arguments = arguments.strip()
+    if arguments.startswith("{"):
+        try:
+            arguments_dict = json.loads(arguments)
+        except json.JSONDecodeError as ex:
+            report_parse_error(ex)
+
+    # otherwise parse it as yaml (which will pickup unquoted strings, numbers, and true/false)
+    # and then create a dict that maps it to the first function argument
+    else:
+        tool_info = next(
+            (
+                tool
+                for tool in tools
+                if tool.name == function and len(tool.parameters.properties) > 0
+            ),
+            None,
+        )
+        if tool_info:
+            param_names = list(tool_info.parameters.properties.keys())
+            try:
+                value = yaml.safe_load(arguments)
+                arguments_dict[param_names[0]] = value
+            except yaml.error.YAMLError:
+                # If the yaml parser fails, we treat it as a string argument.
+                arguments_dict[param_names[0]] = arguments
+
+    # return ToolCall with error payload
+    return ToolCall(
+        id=id,
+        function=function,
+        arguments=arguments_dict,
+        type="function",
+        parse_error=error,
+    )
